Exception handling - delete

+++ src/main/java/com/moisesdias/cursodespringboot/services/UserService.java
@@ -5,6 +5,8 @@ import com.moisesdias.cursodespringboot.repositories.UserRepository;
 import com.moisesdias.cursodespringboot.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.dao.DataIntegrityViolationException;
+import com.moisesdias.cursodespringboot.services.exceptions.DatabaseException;
 
 import java.util.List;
 import java.util.Optional;
@@ -21,7 +23,7 @@ public class UserService {
     public User findById(Long id) {
         Optional<User> obj = repository.findById(id);
 
-        return obj.orElseThrow(() -> new ResourceNotFoundException(id)      );
+        return obj.orElseThrow(() -> new ResourceNotFoundException(id));
     }
 
     public User insert(User obj) {
@@ -29,17 +31,26 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        repository.deleteById(id);
+
+        try {
+            if (repository.existsById(id)) {
+                repository.deleteById(id);
+            } else {
+                throw new ResourceNotFoundException(id);
+            }
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj) {
         User entity = repository.getReferenceById(id);
-        updateDate(entity, obj);
+        updateData(entity, obj);
 
         return repository.save(entity);
     }
 
-    private void updateDate(User entity, User obj) {
+    private void updateData(User entity, User obj) {
         entity.setName(obj.getName());
         entity.setEmail(obj.getEmail());
         entity.setPhone(obj.getPhone());
